requirements.txt
flask==3.0.0
flask-socketio==5.3.6
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.26.3
pillow==10.2.0
python-engineio==4.8.0
python-socketio==5.10.0
eventlet==0.33.3
pyOpenSSL==23.3.0 